Shell sort code in c++
Shell sort is based on insertion sort algorithm. This algorithm avoids large shifts as in case of insertion sort, if the smaller value is to the far right and has to be moved to the far left.
This algorithm uses insertion sort on a widely spread elements, first to sort them and then sorts the less widely spaced elements. This spacing is termed as interval/gap.Here the gap is initially size/2 and then reduces by a factor of 2 till it reaches 0.we can use 3x+1 sequence for spacing as well.Time complexity depends upon the  gap sequence.Here the worst case time complexity is O(n^2).
Algorithm:
Step 1 − Initialize the value of h(gap).
Step 2 − Divide the list into smaller sub-list of equal interval h
Step 3 − Sort these sub-lists using insertion sort
Step 3 − Repeat until complete list is sorted